requirements.txt
python-telegram-bot~=13.9
python-dotenv~=0.19.2
